Update goaltask and copyright info
